+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Changelog
 This file contains highlights of what changes on each version of the [Akismet for PHP](https://github.com/cedx/akismet.php) library.
 
+## Version 2.0.0
+- Breaking change: changed the return type of the `jsonSerialize()` and `toJSON()` methods.
+- Added the `jsonSerialize()` and `toJSON()` methods to the `Client` class.
+
 ## Version 1.2.0
 - Added the `Client::DEBUG_HEADER` constant.